requirements.txt
faiss-cpu==1.8.0
openai==1.34.0
python-dotenv==1.0.1
langchain==0.2.5
langchain-community==0.2.5
langchain-openai==0.1.8
streamlit==1.35.0
unstructured==0.14.6